# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Ineuron Assessment",
    "version": "1.0.0",
    "description": "API for managing items."
  },
  "basePath": "localhost:3000",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/getitems": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get all items",
        "operationId": "getAllItems",
        "responses": {
          "200": {
            "description": "A list of items.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/ResponseData"
              }
            }
          },
          "404": {
            "description": "Item Not found",
            "schema": {
              "items": {
                "type": "object",
                "$ref": "#/definitions/ValidationError"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/items": {
      "post": {
        "tags": [
          "Items"
        ],
        "summary": "Create a new item",
        "operationId": "createItem",
        "parameters": [
          {
            "name": "item",
            "in": "body",
            "description": "The item to create.",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Item"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The created item.",
            "schema": {
              "$ref": "#/definitions/ResponseData"
            }
          },
          "500": {
            "description": "Server error",
            "schema": {
              "$ref": "#/definitions/ValidationError"
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/getitembyId/:id": {
      "get": {
        "tags": [
          "Items"
        ],
        "summary": "Get Items by Id",
        "operationId": "getItemById",
        "parameters": [
          {
            "name": "id",
            "in": "params",
            "description": "A Single Item Details",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Item Detail",
            "schema": {
              "$ref": "#/definitions/ResponseData"
            }
          },
          "404": {
            "description": "Item Not found",
            "schema": {
              "items": {
                "type": "object",
                "$ref": "#/definitions/ValidationError"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/updateitems/:id": {
      "put": {
        "tags": [
          "Items"
        ],
        "summary": "Update Items by Id",
        "operationId": "updateItemById",
        "parameters": [
          {
            "name": "id",
            "in": "params",
            "description": "Update a single Item",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Item"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Updated Item Detail",
            "schema": {
              "$ref": "#/definitions/ResponseData"
            }
          },
          "404": {
            "description": "Item Not found",
            "schema": {
              "items": {
                "type": "object",
                "$ref": "#/definitions/ValidationError"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    },
    "/delitems/:id": {
      "delete": {
        "tags": [
          "Items"
        ],
        "summary": "Delete Items by Id",
        "operationId": "deleteItemById",
        "parameters": [
          {
            "name": "id",
            "in": "params",
            "description": "Delete a single Item",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a single Item",
            "schema": {
              "$ref": "#/definitions/ResponseData"
            }
          },
          "404": {
            "description": "Item Not found",
            "schema": {
              "items": {
                "type": "object",
                "$ref": "#/definitions/ValidationError"
              }
            }
          },
          "default": {
            "description": "Unexpected error",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Item": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "ResponseData": {
        "description": "Successful response",        
          "type": "object",
          "properties": {
            "error": {
              "type": "boolean",
              "default": false
            },
            "message": {
              "type": "string"
            },
            "data": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "name": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "price": {
                    "type": "number",
                    "format": "double"
                  }
              
            }
          }
        }
      }
    },
    "ValidationError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "boolean",
          "default":false
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": null
        }
      }
    },
    "Error": {
      "type": "object",
      "properties": {
        "error": {
          "type": "boolean",
          "default":true
        },
        "message": {
          "type": "string"
        },
        "data": {
          "type": null
        }
      }
    }
  }
}